vyos-kernel: add inotify support for stackable filesystems (overlayfs)
As it stands at the moment, overlayfs doesn't have full support for
inotify, and as such anything that relies on inotify currently has
issues.  The simplest method of demonstrating this is to tail a file
(so tail -f /var/log/messages) and see that it doesn't follow changes
in that file.  This has been reported in a number of places, including
Bug #882147 in Ubuntu.  This patch is based on the version proposed by
Li Jianguo in response to this bug, adding support for inotify in
stackable filesystems.

Bug #303 http://bugzilla.vyos.net/show_bug.cgi?id=303